pass locals to `em.expand` using 4.x compatible method
The 4.x branch expects a locals dictionary whereas 3.x expects keyword
arguments, though it can use a locals dictionary too.

Allegedly on 4.x we should be using the `locals` keyword argument
instead of a positiional argument for maximum future-proofing, but that
is incompatible with 3.x. If it breaks in the future, hopefully 3.x will
be long gone so that argument can be used.

Co-authored-by: Thomas Watson <twatson52@icloud.com>