fix decoding of compound TAO elements
Previously, if any bits were left, the loop would try to decode another
element. However, elements of arrays eligible for TAO must be >= 8 bits,
so the decode would over-run by an element if 7 or fewer bits were left
and the decode would fail due to the length check.

Fix this by only decoding until there are 7 or fewer bits left.

Not triggered by any message in the DSDL repo, but it is easy to create
one which would be incorrectly decoded, e.g.

```
void1 pad
uavcan.equipment.gnss.Fix2 fix
```