avoid sending larger-than-max array lengths over the wire
Avoids creating invalid messages, overflowing the array length, and
having a length that doesn't match the element count.

Slight downside that the user will have no way to know their messages
are invalid (as a larger-than-max length should cause the receive side
to reject the message assuming it does not overflow).