introduce skeleton for table coding system
If the message structure is compatible, the compiler generates a table
which describes the structure. If enabled, an interpreter in libcanard
is then used to follow the table description and encode or decode the
message. This occupies substantially fewer bytes per message as the
table is much smaller than the C function previously used. This
translates to a large reduction in program memory usage, especially if
many different message types need to be coded.

The call to the interpreter (along with a pointer to the message and the
associated table) is inlined whereever e.g. <msg>_encode is called. The
existing function to encode/decode is still generated in case table
coding is not enabled in libcanard or the structure is not compatible.

This commit sets up all the hooks to generate the tables and call the
interpreter functions. However, no tables are generated as the
generation function is a simple stub that always fails.